Clip scratchpad area of ProgField
The ProgField element has a parent rectagle with round corners,
therefore the current implementation uses a smaller rectangle inside it
to clip the scratchpad area inside of it.

It is not the most elegant implementation since this clipping happens
before the elements reach the main rectagle perimeters by a width equal
to the corner radius.