Set limits for dragging scratchpad area
The scrathpad working area, a.k.a. scratchpad area is larger than the
visible size for the user and therefore it is dragable. To prevent the
user from going beyond its maximum size, the scratchpad area vertices
are used as dragging limits, and the maximum dragging happens when they
match the scratchpad visible area corresponding vertices.